Updated dependencies and used latest vue
Updated all dependencies. Made sure that latest version of vue.js is used instead of an outdated one. Used node version 16.14.2. This should fix the "[Vue warn]: Property "$createElement" was accessed during render but is not defined on instance" issue. All dependencies need to be updated to the latest versions to be able to easily make builds in the future.